Host-site time is a double now

+++ src/CustomTextWriter.cc
@@ -55,7 +55,7 @@ CustomTextWriter::~CustomTextWriter()
 }
 
 void
-CustomTextWriter::write(uint numParts, BufferList const& buffers, uint node_offset, float t, const bool testpoints)
+CustomTextWriter::write(uint numParts, BufferList const& buffers, uint node_offset, double t, const bool testpoints)
 {
 	const double4 *pos = buffers.getData<BUFFER_POS_GLOBAL>();
 	const float4 *vel = buffers.getData<BUFFER_VEL>();
