+++ src/UDPWriter.cc
@@ -200,7 +200,7 @@ UDPWriter::~UDPWriter() {
 }
 
 void
-UDPWriter::write(uint numParts, BufferList const& buffers, uint node_offset, float t, const bool testpoints)
+UDPWriter::write(uint numParts, BufferList const& buffers, uint node_offset, double t, const bool testpoints)
 {
 	const double4 *pos = buffers.getData<BUFFER_POS_GLOBAL>();
 	const float4 *vel = buffers.getData<BUFFER_VEL>();
